Moved item-shifting functions from Interface to Utils
